Extract MockLatencyData to separate test helper file
The MockLatencyData class was nested in HttpLatencyLogEnricherTests. Extract it to a dedicated file following extensions repo convention (e.g., CustomHttpLogEnricher.cs, TestHttpLogEnricher.cs) for better test organization.
