+++ test/Libraries/Microsoft.AspNetCore.Diagnostics.Middleware.Tests/Latency/Internal/HttpLatencyLogEnricherTests.cs
@@ -90,50 +90,6 @@ public class HttpLatencyLogEnricherTests
             It.Is<string>(s => s.Contains(ld.SerializedLatencyData) && s.Contains(headerName))), Times.Once);
     }
 
-    private class MockLatencyData
-    {
-        private readonly ArraySegment<Checkpoint> _checkpoints = new(new[]
-        {
-            new Checkpoint("ca", 1, 1000),
-            new Checkpoint("cb", 2, 1000),
-            new Checkpoint("c/c", 3, 1000)
-        });
-
-        private readonly ArraySegment<Measure> _measures = new(new[]
-        {
-            new Measure("m/a", 1),
-            new Measure("mb", 2),
-            new Measure("mc", 3),
-        });
-
-        private readonly ArraySegment<Tag> _tags = new(new[]
-        {
-            new Tag("t/a", "t1"),
-            new Tag("tb", "t/2"),
-            new Tag("tc", "t3")
-        });
-
-        public MockLatencyData()
-        {
-            const int MillisecondsPerSecond = 1000;
-
-            LatencyData = new LatencyData(_tags, _checkpoints, _measures, 20, 1000);
-
-            SerializedLatencyData = string.Format(CultureInfo.InvariantCulture, "{0}/,{1}/,{2}/,{3}/,{4}/,{5}/,{6}",
-                string.Join("/", _tags.Select(a => a.Name.Replace('/', '_'))),
-                string.Join("/", _tags.Select(a => a.Value.Replace('/', '_'))),
-                string.Join("/", _checkpoints.Select(a => a.Name.Replace('/', '_'))),
-                string.Join("/", _checkpoints.Select(a => (long)Math.Round(((double)a.Elapsed / a.Frequency) * MillisecondsPerSecond))),
-                string.Join("/", _measures.Select(a => a.Name.Replace('/', '_'))),
-                string.Join("/", _measures.Select(a => a.Value)),
-                (long)Math.Round(((double)LatencyData.DurationTimestamp / LatencyData.DurationTimestampFrequency) * MillisecondsPerSecond));
-        }
-
-        public string SerializedLatencyData { get; private set; }
-
-        public LatencyData LatencyData { get; private set; }
-    }
-
     private static HttpContext GetHttpContext(ILatencyContext latencyContext, IHeaderDictionary? headers = null)
     {
         var httpContextMock = new DefaultHttpContext();
